refactor(workers): move worker selection logic to separate file
- Created new file worker_selection.go to house worker selection logic
- Moved GetEligibleWorkers, isEligibleRemoteWorker, and RoundRobinIterator to worker_selection.go
- Updated send_work.go to use new exported functions from worker_selection.go
- Renamed newRoundRobinIterator to NewRoundRobinIterator for proper exporting
- Updated imports and function calls in send_work.go to reflect new structure
- Improved code organization and modularity
